factor out common utils
