+++ cypress.json
@@ -2,5 +2,6 @@
   "baseUrl": "http://localhost:2222",
   "projectId": "ma3dkn",
   "viewportWidth": 1500,
-  "blockHosts": ["trackcmp.net", "js.hs-analytics.net", "js.hs-scripts.com"]
+  "blockHosts": ["trackcmp.net", "js.hs-analytics.net", "js.hs-scripts.com"],
+  "testFiles": "**/*_spec.js"
 }
