+++ cypress/integration/examples/blogs/links_spec.js
@@ -1,20 +1,16 @@
 /// <reference types="cypress" />
-const YAML = require('yamljs')
+import { getBlogs, visitBlogsPage } from '../utils'
 
 describe('Blogs', function () {
   let blogs = []
 
   before(() => {
-    cy.readFile('source/_data/blogs.yml')
-    .then(function (yamlString) {
-      blogs = YAML.parse(yamlString)
+    getBlogs().then((list) => {
+      blogs = list
     })
   })
 
-  beforeEach(() => {
-    cy.visit('/examples/media/blogs-media.html')
-    cy.contains('.article-title', 'Blogs').should('be.visible')
-  })
+  beforeEach(visitBlogsPage)
 
   it('lists small links', () => {
     cy.get('.media-small').each((blogEl, i) => {
