+++ cypress/integration/examples/blogs/urls_spec.js
@@ -1,21 +1,16 @@
 /// <reference types="cypress" />
-
-const YAML = require('yamljs')
+import { getBlogs, visitBlogsPage } from '../utils'
 
 describe('Blogs', function () {
   let blogs = []
 
   before(() => {
-    cy.readFile('source/_data/blogs.yml')
-    .then(function (yamlString) {
-      blogs = YAML.parse(yamlString)
+    getBlogs().then((list) => {
+      blogs = list
     })
   })
 
-  beforeEach(() => {
-    cy.visit('/examples/media/blogs-media.html')
-    cy.contains('.article-title', 'Blogs').should('be.visible')
-  })
+  beforeEach(visitBlogsPage)
 
   it('lists large blog urls', () => {
     cy.get('.media-large .media h2 a').each((blogTitle, i) => {
